update transaction checkout

+++ database/migrations/2023_08_24_140944_create_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->uuid('id')->primary();
 
             $table->string("name");
-            $table->float("price");
+            $table->decimal("price", 11, 0);
             $table->longText("description");
             $table->string("tags")->nullable();
             $table->foreignUuid("category_id");
